pyproject.toml
[tool.poetry]
name = "assignment"
version = "0.1.0"
description = ""
authors = ["wlsto <junhu84@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.2"
djangorestframework = "^3.15.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
